Refactor traffic-spectrum.py
Separate the big functions into each little file, whilst
traffic-spectrum is still to be called from the project root.

This makes it easier when trying to do things such as using different
down-samplers, or using different spectrogram generators.
